SpMixing: First version
This commit allows to mix storage pools arbitrarily. This is done
as follows:
* All satellites report their storage pool's extent size to the
controller
* The controller runs a check whether or not storage pools are mixed.
There are a few factors that can lead to a positive check:
  * Different thin-types (thick vs thin allocation)
  * Different extent sizes
That means, that even two LVM_THIN storage pools with different
extent sizes are considered as "mixed storage pools".
* If a resource-definition is marked as "mixed storage pools",
a DRBD-full-sync is always performed (no initial partial sync).
* Additionally, the least common multiple of all participating
extent sizes is stored on the volume-definiton. This value
is taken by all satellites when creating a new LV to round up to.

This way there are now potentially two differences in how a new
resource is created from the satellite's perspective:
1) The new resource will not mark itself as UpToDate, even if
it is backed by a thin storage pool
2) The new resource is either as large or larger (but not smaller)
than the backing discs of existing peers.

These conditions are necessary because of the following reasons:
* Violating the first condition will result in a scenario where the
new peer has a larger disc AND regards itself as UpToDate. Therefore
it will reject the connection to a peer with "too small disc".
* Violating the second condition results in the opposite error, that
the new (local) peer has a smaller disc than existing peers and will
therefore not be able to receive all data.

Currently this feature only works for manually created resource,
no autoplacer, etc..