Snap,Rollback: Fixed unrecoverable state with failed rollback and deleting snapshot
If a snapshot is lost on LVM level, a "snapshot rollback" will obviously fail.
That rollback will set the property "RollbackTarget" on the resources for the
rollback-attempt. If a rollback fails, "aborting" a rollback was not an option
(until now), so the property stayed on the resources.

If afterwards the snapshot was also marked for deletion, this situation was no
longer recoverable, since all satellites will first attempt to delete the snapshot
(which might run in a noop) and afterwards try to rollback the (just deleted)
snapshot. This will fail forever and the user cannot recover from this without
manually modifying the database.

This patch changes this problem by introducing a new callback API from the
satellite to the controller stating the success/fail of every individual rollback.

If all satellites succeeded the rollback, the "RollbackTarget" property is removed
(as usual) and the rollback operation is a success.

If all satellites fail the rollback, the same procedure happens as if the rollback
was a success - with the only difference that the client does get ERROR messages.
The "RollbackTarget" is deleted.

If some satellites fail and some succeeded (this should only be possible using
DRBD), then the "RollbackTarget" property is also deleted (to avoid the bug described
earlier), but additionall all failed resources will be marked to create new DRBD
metadata for themselves so they can sync up from the peers that succeeded the rollback.
Recreating the DRBD metadata is critical, since after a snapshot was created LINSTOR
did not tell DRBD to create a new GID, which means that without new metadata all
satellites (failed as well as succeeded) will show themselves as UpToDate with the
same GID (connected and apparently all happy) but with different backing data. DRBD
also has no idea which blocks are different.

A different approach would be to run an invalidate right after the "drbdadm adjust"
but that also has other problems. Recreating the metadata might "unnecessarily" full-
sync all data, but this is surely the most robust way to ensure all peers have the
same data again in the end.