Backup: Add force-restore
Currently --force-restore has the following criteria to work:

* has to be explicitly set by the user
  currently supported operations are:
  * backup restore (from s3)
  * backup ship    (linstor to linstor)
  * backup schedule enable (also only feasible for linstor-to-linstor
    since we do not have a "scheduled restore from s3")
* the target resource-definition must not be in use
* the target resource-definition has at most 1 resource deployed
* the target resource-definition allows for force-restore
  This means that the property BackupShipping/AllowForceRestore
  is NOT set to False on RD, RG or C level. The property is True
  by default.

If force-restore is possible and no resource exists on the target-RD,
the command behaves exactly the same as without the --force-restore.
If force-restore is possible and one resource exists on the target-RD,
that resource is deleted (possible loss of data!) and the command continues
with the usual restore procedure.
If more then one resource exists, force-restore is not possible and will
cause a warning and perform a --download-only