NodeLost: prevent race condition
When a node goes offline or is deleted, its corresponding backup
queues need to be cleaned up. Previously, node lost relied on the
cleanup that happens when the ctrl loses connection to the node
for this. Unfortunately this allows for a race to happen:

node goes offline -> ctrl notices node offline -> cleanup happens
                                               ^
                                        node lost command

When the node lost command is processed at the indicated time,
the node is already deleted when the cleanup should happen, leading
to an AccessToDeletedDataException. This is now fixed by ensuring
that node lost calls the cleanup on its own, and additionally
adding an isDeleted-check to the cleanup method.