Drbd,forget-peer: Rework deleting resource
Diskless peers can also call del-peer, since that only calls "drbdsetup ...".
"drbdadm forget-peer ..." would also call "drbdmeta ..." which is a bad idea
for both, diskless and skipDisk situations.

If we have a skipDisk situation but we want to delete a peer, the "forget-peer"
is quite important to execute, but that cannot be done if we have to assume
that our disk is broken (maybe just temporarily?).
Therefore this patch stores the node-id of the peer as a property (similar
as backup restore already does), and runs the "forget-peer" as soon as the
disk is available again (and deletes the resource-properties again)