snapshots: fix string builder exception
When updateDescriptionAndJoinedRscNames() is called with no snapshot
to be displayed, the manual string join implementation fails: it tries
to set the length of the string builder to -1.

To fix, we use String.join() to construct the string instead: this has
a much better fallback behaviour if the list to join on is empty.