Backup,restore: Recalculate missing AllocationGranularity
If a pre-v1.26.0 LINSTOR created a backup, that snapshot will not contain the
AllocationGranularity property. If the source cluster and the target cluster
have different extent sizes, that missing property will cause problems, showing
as the DRBD error "peer disk too small".
v1.26.0 deals with those errors with the mixed storage pool feature, but that
feature depends on the AllocationGranularity property. If that property is still
missing (due to being uploaded with an older version of LINSTOR), the same
issues will arise.

This patch tries to estimate the allocation granularity, as described in the
Javadoc. The goal of the calculation is not the be precise, but to be
"good enough" so that new peers can join without running in the "peer disk too
small" issue.