evacuate/migrate: wait until resource is not InUse before removal
When evacuating or migrating an InUse resources, the existing implementation
waited until the resource was UpToDate on the new peer, then removed the
resource.

However, since the resource is still InUse on the migrated from node, DRBD
will immediately outdate the other resource upon (graceful) disconnect,
while also not being able to shut down the resource on the migrated from node.

To fix this, we introduce a second state check, that ensures the resource
is only downed once no longer in use.