LvmCommands: Remove the trailing "accept everything else" from filter
Copied from the Javadoc:
--------
DO NOT add "'a|.*|'" at the end of the filter list. If a block device has multiple paths (symlinks) the device is
accepted if ANY of the paths matches. Only if all paths either have a reject first of are undecided, the device
is rejected.

We had cases where we rejected "/dev/drbd" but LVM still found (the suspended) DRBD device via
/dev/block/147:1000. The reason LVM took that path is because we allowed "everything else" (using "'a|.*|'")
--------