Luks: Allow deletion without master passphrase
This fix consists of two parts:

The satellite was changed that it now tries Luks operations for which
no luks-key is required (check if device exists and "cryptsetup close"
for example). Only operations that actually need the luks-key will now
complain (via a thrown exception) if the key is missing.

The controller no longer sets the ignore reason "LUKS no key" if the
resource in question has the DELETE or the INACTIVE flag set. This way
the controller allows the satellite to delete the luks resource even
if the master passphrase was not entered since the last controller-
restart.