SysFs: Reworked processing
The following points were addressed:

* SysFsHandler no longer throws an ImplementationError if a
StorageException occurs. We had cases where /sys/fs/cgroup/blkio was
missing, which caused an IOException when trying to access the device.
That IOException is converted into a StorageException, which was thrown
as an ImplementationError, aborting the DeviceManager run completely.
This prevented DevMgr from answering the controller (no success but also
no failure), making the controller waiting (sometimes indefintely long).

* SysFsHandler now operates on single resources instead of waiting
for a list of resources. This allows us to add the resource name in the
warning message.

* SysFsHandler returns a warning (visible for client commands) in case of
a StorageException.

* SysFsHandler's StorageException now also include the underlying
IoException for a more detailed error cause.