Rs-discard-gran: Deleted old prop handler
The old property handler took the value of the thin pool's granularity
in case of LVM_THIN (or 64k as default) or 8k in case of ZFS and ZFS_THIN.

The problem with this was, that if the storage pool's granularity is larger
than 1M (which DRBD's maximum), the property is still set and sent to the
satellites. Every satellite, however, performs a check if that value is
valid or not. If the storage pool for example now has a 2M granularity
(chunk-size in LVM speak), those 2M are copied into the VD property, but
will fail the satellites validity-check, causing annoying Problem-reports
(those also create ErrorReport-*.log files, but unlike actual errors those
do not show up in "linstor err list")

Deleting this old mechanism was already enough, since the new version was
already in place. The new mechanism handles that 1M limit correctly, but does
so in a later flux-step, so that the satellite already received the first (and
wrong) update.