FullSync: Improve error message when ext tool is missing
We have seen cases where an OS upgrade unintentionally downgraded
DRBD to 8.4 (upstream). This resulted in hard to connect ErrorReports
about "NullPointerExceptions" while verifying properties. The connection
there is that the satellite rebuilds the allowed list of properties
while the controller is connecting. An 8.4 DRBD simply does not contain
some of those information (or does not support xml-help at all, not sure).
Anyways, that results that the lookup tables are empty. The further
processing causes then an NPE.