AbsStorProvider: Delete unnecessary volumesLut
That LookUpTable might be a bit faster than the current implementation, but
requires the volume to be processed simultaneously with all the snapshots.
Since we want to decouple processing of resources and snapshots, this LUT
will no longer be needed.

If for some reason this commit needs to be reverted, we still can pass the
resource to each snapshot-processing, which should achieve the same effect.