DevHandler: Split processing of resources and snapshots
If a resource or a snapshot ran into an error, the processing of both was
aborted. That especially means, if a resource has a broken snapshot which is in
DELETE state (i.e. get processed first in the AbsStorageProvider), neither
the resource, nor other (unrelated) snapshots are processed at all.

That leads to the following scenario:
* we start with a resource, called "rsc1"
* create a snapshot "snap1" of "rsc1"
* restore "snap1" of "rsc" into a new resource "rsc2", on ZFS. ZFS has a parent-
child dependency in this case, so "rsc1@snap1" cannot be deleted while "rsc2"
exists.
* being unaware (or by mistake) still try to delete "snap1" of "rsc1". This will
fail. This patch does *NOT* attempt to fix this. This is an expected failure
due to ZFS's behavior.
* create a second snapshot "snap2" from "rsc1". This however also failed before
this patch, as mentioned earlier (all other snapshots and the resource will
not be properly processed if a snapshot fails).

With this current patch, snapshots and resources are processed independently of
each other, allowing the last step of the above scenario to succeed.
(again, just to make sure, "snap1" of "rsc1" will still remain in DELETE state
with a persistent error, this is not what this patch tried to fix).