K8s,SpaceHistory: Fix incorrect dates
Before we unified the three database drivers of the SpaceHistory table,
the K8s version accidentally stored the date with the current time in the
database.
SQL was told to only store DATE (not DATETIME), so SQL auto-corrected LINSTOR's
mistake. For ETCD we used our own String.format that only included year, month
and day, so we do not need migrations for those DB types.

This migration is not perfect, since the previous K8s driver also allowed to have
multiple entries per day (with different times). This migration just takes the first
entry it finds for a given day, not strictly the latest entry.

These incorrect date-format lead to the following issue:
Let us assume that we have a date 01-01-2024 01:15:00 stored in the database (note that
the time should be 00:00:00). When we load it from the Database, the date get sanitized
in the holder-class that wraps the entry date with the capacity. That means, although
K8s stored the date(time) with 01:15:00, LINSTOR processes that entry object as if it
was 00:00:00.
When an entry gets too old, LINSTOR tries to delete it. Doing so, LINSTOR is convinced
that the entry (as it is) also exists in K8s. However, the actual deletion will fail, since
the derived key from the entry "01-01-2024 00:00:00" simply differs from the original key
that was derived from "01-01-2024 01:15:00". So the deletion fails with "unknown key" or
something similar.