AutoSnapTask: Fix rescheduling issue
Some methods changed the AutoSnapshotConfig.runEveryInMs and/or the
AutoSnapshotConfig.nextRerunAt. Those variables (as mentioned in the
java comment) are part of the compareTo method, changing the ordering.
If that ordering changed without properly updating the containing sorted
collection, the .add, .remove, .contains and other methods might not work
properly.

To be more precise, this  means that even if we iterate in a loop over
the sorted collection, pick an element, and call (after the loop) the
.remove method of the sorted collection, the element might still be
contained in the collection.
Or the .add method might add the modified element a second time (in its
"correct" order/position)