Rsc,del: Rerun tiebreaker-check after migration
In a scenario where we had only a single DRBD resource (no peers)
and evacuate the node with the single peer, the following happened:
1) on a second peer a diskful resource is created which will start
syncing up soon
2) since we have now 2 diskul peers, on a third peer a tiebreaker
resource is created
3) once the sync is complete, peer 1 gets deleted.

This last deletion missed the tiebreaker-check, so we were left
with a diskful resource on peer 2 and a tiebreaker on peer 3. A
tiebreaker does not make sense with only a single diskful peer.

The main change of this patch is in CtrlRscDeleteApiHelper which
now simply re-runs the auto-helpers. Since the auto-helpers need
ResponseContext, the other classes had to be modified as well
(new parameter for the methods, etc...)