ExactSize: Add property to render the size entry
This property should only be used for migration purposes from LINSTOR-external
devices into LINSTOR where the device size of the source and the (LINSTOR-)
target has to match.
After that operation it is strongly advices to delete this property again.

"DrbdOptions/ExactSize" property can only be enabled ("True" value) if the
given resource-definition has no resources.

While the property is enabled, the volume definitions cannot be resized.

The property can be disabled at any time.