docker: fix udev sync in containers
Next operator release mounts /run/udev into the container because that is
needed to get ZFS working. However, this had the side effect of breaking LVM.

LVM is by default checking for the existence of this directory to determine if
it should wait for udev events. This is however broken inside containers, so
the LVM commands all time out, breaking LINSTOR with weird errors all around.

To fix this, disable the "udev_sync", so LVM no longer waits on udev events.