SkipDisk: ensure all layers below DRBD are ignored
If a disk has a failure, which results in SkipDisk being set, an
lvm list will no longer show said disk. That leads to linstor
trying to create that missing disk, which will lead to errors since
the disk is broken. To avoid these exceptions, the storage layer
will now be ignored, using the ignore-reason feature.