Autoplacer: Reworked SelectionManager rollback mechanism
Before this patch, the selection was stored in a set (like selectedNodes,
selectedSp, etc...) and from these sets the current fixed and consumed values
for replicas-on-same and -different were recalculated during an unselect.

Now the SelectionManager stores an unmodifiable state, which can simply be
thrown away during an unselect and the topmost from a stack of states used.