Autoplacer: Added x-replicas-on-different
This patch does not really touch the existing replicasOnDifferentMap but rather
adds a new xReplicasOnDifferentMap to the autoplacer.
This new map only stores the property key to the allowed number per value.

This feature is indented for following setups:
Lets say we have multiple datacenters with each having multiple nodes. We want
LINSTOR to create 2 replicas per datacenter. We have already set
"Aux/DC = ..." properties on each node accordingly. The autoplacer call should
look like:
 --x-replicas-on-different 2 DC --place-count 5
In this case we expect three datacenters to be selected, two having 2 nodes
(i.e. fully saturating the X in --x-replicas-on-different) and the third
datacenter having only 1 replica.

A usual "--replicas-on-differnt DC" silently assumes a
"--x-replicas-on-different 1 DC"