clone: open/close interface, find common clone-strat
This implements a open/close interface for layers on the satellite,
and an algorithm that is used to find the best common clone-strategy.

There are quite a few cases which we have to distinguish. A few examples:
DRBD,LVM -> DRBD,LVM: We can ignore DRBD completely and just clone on LVM level
    using dd.
DRBD,LVM_THIN -> DRBD,LVM_THIN: Again, DRBD can be ignored and we can use a
    snapshot-based clone (since we are using LVM_THIN) instead of dd.
DRBD,STORAGE (any) -> DRBD,LUKS,STORAGE: Although DRBD can be ignored again, we
    still need to use dd (regardless if both STORAGEs have the same provider or
    not), since the data needs to be written through the LUKS layer.
DRBD,STORAGE -> STORAGE: In this case, DRBD must not be ignored, since we want
    to get rid of DRBD's metadata. Technically we could still use faster-than-dd
    methods in case we have DRBD with external metadata (since we just want to
    get rid of them anyways)
STORAGE -> DRBD,STORAGE: Regardless of internal or external DRBD metadata, we
    must not ignore DRBD in order to prevent the permanent usage of
    "ForceInitialSyncPermanent".
DRBD(ext-md),STORAGE -> DRBD(internal-md),STORAGE: We must not ignore DRBD here,
    regardless of the actual provider kinds.