Backup: drop the trap SIGTERM () & wait $! dance from shell snippets
since v1.24.2 a4fb7a78c (Backup: Replace setsid -w with timeout 0, 2023-08-24),
we use coreutils `timeout` to propagate signals to the process group.

We no longer need to use a signal trap in the shell to try to do the same.
And if we don't need that, we don't need the extra subshell in the background
anymore. And that again means we don't need the extra file descriptor
redirection dance either and can use stdin as is.

    exec 7<&0 0</dev/null;
    set -o pipefail;
    trap 'kill -HUP 0' SIGTERM;
    ( exec 0<&7 7<&-; funky | pipeline ) & wait $!

becomes just

    set -o pipefail;
    funky | pipeline