Backups: ensure abort works if triggered early
When aborting an l2l-shipment, it is possible to issue the abort
while the backup is in a weird state where the snap-dfn has the
SHIPPING flag set already, but the snap itself is missing the
BACKUP_SOURCE flag. Previously this state did not get registered
as a backup in need of aborting, which can lead to unabortable
as well as undeleteable snapshots should an error cause the flags
to get stuck in that state. Now the abort recognizes these symptoms
and correctly aborts such shipments.