BackupShipping: suppress receive-thread errorReports while aborting shipment
When aborting an l2l-shipment on the src-ctrl, the receive-thread on
the dst-cluster might end due to errors (stream ended 'unexpectedly').
These errors are expected and therefore do not need to generate an
error-report. These reports are prevented by having the src-ctrl send
a PrepareAbortRequest to the dst-ctrl which makes the dst-ctrl set
the new PREPARE_SHIPPING_ABORT flag, which in turn stops the
receive-thread from creating error-reports while the sending-thread
is aborted. It also prevents the receive from starting if it hasn't
yet done so (since the shipment is being aborted, there's no need to
start part of it).