Spawn: Improve diskless flag calculation for diskless-on-remaining
The bug this patch tries to fix is the following:
* spawn a new resource with --diskless-on-remaining (at least with 1 diskless)
* toggle-disk the diskless resouce.

That toggle-disk fails because toggle-disk only removes the DRBD_DISKLESS flag
(since toggle-disk only makes sense with DRBD) but later when the
CtrlStorPoolResolveHelper tries to figure out if the resource is diskless or
not, it checks for both flags (DRBD_DISKLESS and NVME_INITIATOR). Since the
NVME_INITIATOR flag is still set (from the spawn command) the
CtrlStorPoolResolveHelper complains that it found a diskful storage pool (which
would be fine in the above test-case) but the resource is supposed to be
diskless (due to the still set NVME_INITIATOR).

Removing the NVME_INITIATOR flag during the toggle-disk would have been a bad
idea, since it does not make sense to allow the user to change an NVMe initiator
into an NVMe target. Also, that would prevent the use-case that we legitimately
have an NVME_INITIATOR resource, which we want to (DRBD-) toggle-disk twice.