Backups: ensure l2l aborts work in both directions
l2l-aborts now happen in several steps: First, the cluster where
the command was issued figures out which snapDfns need aborting.
It then tells the other cluster about those so that the abort
doesn't trigger unneeded (because they are expected) error-reports.
After that is done, the actual abort finally happens.