Backup,restore: Use correct property before recalculating AllocationGranularity
The properties differ only in case:
Storage/internal/AllocationGranularity   (correct, new/current)
Storage/Internal/AllocationGranularity   (incorrect, old)

This bug caused the following error:
* Create a resource with 64MiB (or any other value than the default 4MB) physical extent size
* Create a backup from this resource
* Restore this backup to a default 4MB physical extent size storage pool
* Adding a new peer to the restored resource will error out with "peer's disk is too small".

The reason behind this is that the 64MB AllocationGranularity was correctly set on SnapVlmDfn
and also on VlmDfn, but was not found when checking if a recalculation is required.
Therefore the recalcuation is perfomed, even if not needed.
The problem with that is that the next few lines after the changes of this commit, the "bds"
variable will be the exact same as the "minimumSize" variable, making "diff == 0". The two
variables will be the same because the "calculateSize" method also checks for the VlmDfn
property (the correct one) so it uses the correct 64MiB for the calculation.
Having a "diff == 0" will override (the correct) 64MiB of SnapVlmDfn with 1KiB. This
will be again overridden on the VlmDfn later when the first volume/resource is restored,
but only to the value that is now available, i.e. from the storage pool, i.e. to 4MiB.

The next peer that will be created only rounds up to those 4MiB, which will simply
create a too small device.