Autoplacer,Strats: Don't override defaults when strategies namespace exists
A default value should only be overridden if the corresponding property exists.
The problem with the previous implementation was, that as soon as the namespace
exists, all stretegies are loaded. If the property is missing for a specific
strategy, 1.0 is used as a default, instead of the actual default (which is 0.0)