Autoplacer,MinRscCountStrat: Change default weight to 0.00001
This should have little to no effect to existing setups when placing diskful
resources. However, this change does have an impact when choosing diskless
resources, like for a tiebreaker.

Before this change all storage pools were rated the same, and therefore the
alphanumerically first node was always chosen as a tiebreaker.
With this change, the resource-count will have a minimal impact on the choice,
but since the free-space of diskless storage pools are still exactly the same,
that minimal impact will distribute the tiebreaker better across the available
nodes.