Refactored ReadOnlyProps
The old class ReadOnlyProps got renamed into ReadOnlyPropsImpl and a new
interface "ReadOnlyProps" was created. All getter-methods were moved from
the Props interface into the new ReadOnlyProps interface, and the Props
interface extends the new ReadOnlyProps interface.

The only inconsistentcy is the getNamespace method, which returns a
Optional<Props> type, even in ReadOnlyProps, instead of Optional<ReadOnlyProps>.
The reason for this is that the Props interface (which extends ReadOnlyProps)
must not modify the Optional's generic parameter.

The reason for this is that implementations can now choose to explicitly
store / use the ReadOnlyProps interface, which simply does not contain
setter methods at all. This will cause compiler errors instead of Runtime
errors in case a ReadOnlyPropImpl is used as an implementation.

This commit got quite large, since we tried to replace all usages of Props
with ReadOnlyProps where feasible.