Refactor date-formatting to be thread-safe
Since java.text.DateFormat is not thread-safe, and linstor works
with multiple threads, using java.time.format.DateTimeFormatter
instead is preferable. This means that java.util.Date should also
be replaced with corresponding java.time.* classes where possible.

Generally LocalDateTime is a safe bet, but if a date needs to be
the same no matter where the user is located, ZonedDateTime (most
likely with ZoneOffset.UTC) should be used.

The Strings resulting from a .format stayed mostly the same, only
slight changes were made in some places where the result is not
parsed later on to reduce the number of different formats used.
The common date-formats are found in TimeUtils, as well as some
helpful methods to convert to and from milliseconds (since there
is no easy way to do so with LocalDateTime...).