Introduce StorPoolInfo and NodeInfo
These two interfaces are implemented by StorPool and Node class respectively.
The purpose of these interfaces is to allow a read-only implementation of these
classes, so that the satellite can improve its concurrent processing of requests
(which will be done in a future commit).

This commit replaces on the satellite's layer-processing classes the usage of
StorPools with StorPoolInfo which are used by getSpaceInfo method as well as
some serialization utility in the server project.

Noteworthy is also that some Storage-provider's checkConfig content got moved
into their update method, since the checkConfig is now StorPoolInfo based which
is not allowed to update (aka modify) the StorPool itself.