RequestThinFreeSpace: Allow concurrent execution with devMgrRun
When the device-manager starts its run it takes basically all write locks to
ensure that the data it is working with is not modified. This in itself makes
sense.

However, the controller have multiple sources that will trigger a
RequestThinFreeSpace API call (on every satellite), like
"resource / volume / storage-pool list", "metrics" and others. Before this
patch, fetching free space also needed a few readlocks since storage pools and
their properties need to be accessed. This caused the issue that while the
device manager is working on the resources, which can take quite some time,
especially during startup where all resources need to be checked/adjusted,
fetching free spaces are blocked, and therefore also the client's request.

The core idea of this patch is to copy the required data of the storage pools
into a new data structure StltReadOnlyInfo. The device manager creates a new
instance on every run and atomically updates the class-variable which is then
used by the RequestThinFreeSpace API call. This atomic update should allow us a
lockless approach for this issue.

StltReadOnlyInfo contains two additional classes, ReadOnlyStorPool and
ReadOnlyNode. These two new (sub-)classes also implement the recently introduced
interfaces StorPoolInfo and NodeInfo respectively.

If for some reason this patch does not work or causes issues that requires a
revert, we might want to introducing a new lock which is taken by the
device-manager at the beginning, update this new data-structure and release the
new lock again. The RequestThinFreeSpace API would only need to wait for this
new lock (which should be a much shorter wait-time than for the usual
LINSTOR-locks).