systemd: allow satellite to handle shutdown
In the default settings, systemd will send the SIGTERM signal to all processes
started in a service once you run "systemctl stop ...". This may cause some
random error reports if the Satellite is currently waiting for a command to
finish.

Instead, use KillMode=mixed. As per man systemd.kill, this will send the
initial SIGTERM only to the main processes (i.e. the satellite), the later
SIGKILL is still send to all processes assuming the satellite did not
properly shut down in time.