Lvm,Cache: Add time-based cache for lvs,vgs commands
Default cache time is 10 seconds. All caches are automaically invalidated
if any changing LVM command is executed (like lvcreate, lvresize, etc...).

Note: LvmProvider also needs to recache vgs cache (not just lvs cache) after
a modifying LVM command to properly query the free space information for the
storage pool. For LvmThin recaching lvs cache should be enough.

The property "StorDriver/SizesCacheTime" can be used to (on controller and
node level) to change the caching time. Value is in milliseconds.
When the value is applied, all old entries are automatically purged. This means
that setting the property briefly to 0 and then deleting it again can
effectively be used to completely clear the caches.