LayerIgnoreReasons: Changed nullable String to Set
Instead of an optional single ignore reasons resource-layer-data now
can have a set of ignore reasons.

If any of these ignore reasons wants to prevent the execution of
the specific layer, the execution is prevented

Since we have a set of ignore reasons now, the LayerIgnoreReason.NONE
is no longer needed.