DevHandler: Implement ignoreReason's preventOnDelete
Some ignore reasons (like "DRBD skip-disk") want to prevent execution
of some layers under normal circumstances, but not in others.
One such case is the deletion of a resource.

Imagine a resource that has the SkipDisk property set (and therefore
also the same ignore reason for all layers but DRBD). If this resource
is deleted, since the DRBD layer is not ignored it will "down" the
DRBD resource. However, the storage layer is currently ignored so it
will leave DRBD's backing device untouched. The satellite still reports
the controller a success which leads to the deletion of the LINSTOR-resource
from the database.

This patch tries to fix this issue so that the skip-disk ignore reason
can be ignored during a deletion.

However, if the skip-disk was enabled because the entire disk is broken
and the lvremove will throw IO errors, we do not want to report these
errors. Just "try once, and if that fails, just give up". This however
still has the risk of letting an LV get orphaned.