RG,modify: eagerly subscribe to flux-list
Eagerly subscribing to the list of resource-definition-update-fluxes
causes all updates to be executed even if some of them fails.
Chaning this list of fluxes together using concatWith would abort the
rest of the chain if an error occurs.

However, as the TODO states, we should implement a new ctrl->stlt API
to update a RG directly instead of updating every single of its RDs.