FsWatch: Fix usage
Instead of all providers creating their own FileSystemWatch instance, they rather
now let the same FSW inject. The problem with the former approach was that the FSW
instances simply were never started (it would also have been hard to properly stop
them again).
Since they were never started, the registered "notify me when device X shows up"
never triggered. If the device existed while the callback was registered, the
callback was immediately triggered by an additional mechanism. If the device did
not exist at that time, the callback was never notified; rather, yet another
mechanism performed a check after the wait time was over to see if the device now
exists or not.

That means that LINSTOR in many cases needlessly waited about 5 seconds, even if
the device showed up within a few milliseconds (but a bit too late to be detected
during registering the callback).