Backup,ship,restore,schedule enable: Make target resource group configurable
The new dst_rsc_grp JSON field allows the user to specify the resource-group on the
target-cluster.

If the target RD did not exist before it will be created with the target RG.
If the target RD already exists and the data is not forcefully reset to the content of
the source RD, the RG of the target RD will only be changed if the force_mv_rsc_grp is
set to true.