Autoplacer: Fix possible x-repl-on-diff violation
Simple reproducer:

* 3 nodes each in different Aux/dc
* RG with --x-replicas-on-different dc 2 --place-count 2
  * spawning with this RG should and will fail
* Reduce RG to --place-count 1
  * spawn will and should succeed
* Revert change: Increase RG's --place-count 2
  * autoplacement should have failed but suceeds prior this patch

The issue was that only the list of --replicas-on-different reduced
the "allowed maximum value count" (which is by default 1 unless
--x-replicas-on-different states otherwise). If only
--x-replicas-on-different was used without --replicas-on-different
the SelectionManger simply did not check which values are already
"consumed", allowing a new group for the --x-replicas-on-different
values.

This patch fixes the issue by combining the keys from both,
--replicas-on-different and --x-replicas-on-different and therefore
counting the value count in both cases.