Storage: Refactored usage of asSnapLvIdentifier
The helper-methods were deleted and now every provider has to implement its own
version of asSnapLvIdentifier(LAYER_SNAP_DATA).

The main reason for this is that ZFS not got a new (optional) rscFormatSuffix
part in the snapshot identifier format. Since this change is only related to
ZFS, there is no need to bother other providers with such an optional parameter.