Props: Split Snap- from Resource-properties
Until now snapshots only had a single property container which captured the resource's properties.
(This is also true for snapshot-definitions, -volume and -volume-definitions respectively, but for
simplicity we will only talk about snapshots here).
However, with features like backup shipping, EBS and others there are already a few propertie that
need to be set on snapshot level which should not be copied into a new resource when restored or
rolled back.

Because of this, this patch now introduces a new set of properties for snapshots (and -*).

The still existing "props" field in the corresponding JSON APIs is now deprecated in favor of
newly introduced "snapshot_props" and "resource_props". The deprecated "props" map still contains
the squashed properties from resource_props + snapshot_props (in this order, in case there is an
overlap).

Database migration splits the properties into the corrsponding instance-names.
Instance-names of properties have in general be shortened to be more unified with other usages
(and easier to read).

Since existing S3 meta-files cannot be modified during this migration, the S3 pojo class has a
simple converting mechanism. If an old S3 meta-file is parsed using Jackson, the old "props"
entries will be split into "snapProps" and "rscProps" automatically. The updated pojo no longer
contains a "props" map, only the two new "snapProps" and "rscProps" version. This should
allow us to read the old and the new format but only allow us to write the new format.

Of course this means that an S3 backup created with a new LINSTOR version will not be downloadable
from an old LINSTOR version.